requirements.txt
torchvision
IPython

notebook
pytest
pycodestyle
matplotlib
numpy
scipy
scikit-learn
neural-pipeline-search==0.11.1
xgboost
pyyaml
pickle